Update inlets-pro to 0.6.0
This updates the inlets-pro version downloaded on the exit node to the
latest 0.6.0 version.

This has been tested by running an inlets-pro exit node and trying to
connect with an old license (no claims), which gave the error not valid
for inlets-pro, then changing to a correct inlets-pro license. This new
license worked as expected

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>